Fix the normalizing vector in compute-lid.py
Now the distances in the MLE estimator are normalized to the k-th
neighbor, as they should. Before this change, distances were always
normalized to the last vector in the collection of distances, which in
our particular setup was 100. Therefore estimates for k=100 (the ones
used in the paper) are OK, but estimates with a different k are not.